backend/requirements.txt
# Core DSPy and AI dependencies
dspy-ai>=3.0.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Ollama support for local development
ollama>=0.1.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data handling
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# ML and optimization
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0

# Database and caching
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Monitoring and logging
mlflow>=2.8.0
wandb>=0.15.0
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Environment and config
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Additional AI/ML tools
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Async support
aiofiles>=23.0.0
httpx>=0.25.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
